create an index.test.js file so that our testable app tests file swapping
