+++ test/CavyTester/src/scenarios/index.js
@@ -1,0 +1,33 @@
+// Import new scenarios here:
+import * as containsText from './containsText';
+import * as containsTextNumerical from './containsTextNumerical';
+import * as containsTextUnwrapped from './containsTextUnwrapped';
+import * as pressClassComponent from './pressClassComponent';
+import * as pressFunctionComponent from './pressFunctionComponent';
+import * as exists from './exists';
+import * as fillIn from './fillIn';
+import * as findComponent from './findComponent';
+import * as focus from './focus';
+import * as notExists from './notExists';
+import * as taggingTest from './taggingTest';
+import * as useRef from './useRef';
+import * as createRef from './createRef';
+import * as callbackRef from './callbackRef';
+
+// Add new scenarios here:
+export default scenarios = [
+  containsText,
+  containsTextNumerical,
+  containsTextUnwrapped,
+  pressClassComponent,
+  pressFunctionComponent,
+  exists,
+  fillIn,
+  findComponent,
+  focus,
+  notExists,
+  taggingTest,
+  useRef,
+  createRef,
+  callbackRef
+];
